refactor: moved roles into separate directory
as requested in the PR his file is now in a fixtures directory specifically for json files
